+++ PACECloudSDK/AppKit/AppManagement/MessageHandling/App+MessageHandling+Location.swift
@@ -8,6 +8,7 @@
 import CoreLocation
 import Foundation
 
+// MARK: - Verify location
 extension App {
     func handleVerifyLocation(with request: API.Communication.VerifyLocationRequest, completion: @escaping (API.Communication.VerifyLocationResult) -> Void) {
         let locationToVerify = CLLocation(latitude: request.lat, longitude: request.lon)
@@ -42,5 +43,33 @@ extension App {
 
     private func respondToVerifyLocation(isInRange: Bool, accuracy: CLLocationAccuracy?, completion: @escaping (API.Communication.VerifyLocationResult) -> Void) {
         completion(.init(.init(response: .init(verified: isInRange, accuracy: accuracy))))
+    }
+}
+
+// MARK: - Get location
+extension App {
+    func handleGetLocation(completion: @escaping (API.Communication.GetLocationResult) -> Void) {
+        oneTimeLocationProvider.requestLocation(useLastKnownLocationIfViable: true) { [weak self] userLocation in
+            guard let userLocation = userLocation else {
+                self?.passGetLocationToClient(completion: completion)
+                return
+            }
+            self?.respondToGetLocation(userLocation: userLocation, completion: completion)
+        }
+    }
+
+    private func passGetLocationToClient(completion: @escaping (API.Communication.GetLocationResult) -> Void) {
+        AppKit.shared.notifyCurrentLocation { [weak self] userLocation in
+            self?.respondToGetLocation(userLocation: userLocation, completion: completion)
+        }
+    }
+
+    private func respondToGetLocation(userLocation: CLLocation?, completion: @escaping (API.Communication.GetLocationResult) -> Void) {
+        if let location = userLocation {
+            let coordinates = location.coordinate
+            completion(.init(.init(response: .init(lat: coordinates.latitude, lon: coordinates.longitude, accuracy: location.horizontalAccuracy))))
+        } else {
+            completion(.init(.init(statusCode: .notFound, response: .init(message: "Couldn't retrieve the user's location."))))
+        }
     }
 }